Formatting fixes
