+++ README.md
@@ -151,7 +151,7 @@ repositories {
 }
 
 dependencies {
-    implementation("com.cognifide.gradle:aem-plugin:6.0.2")
+    implementation("com.cognifide.gradle:aem-plugin:7.0.0-beta")
 }
 ```
 
@@ -560,7 +560,7 @@ Then file at path *build/aem/aemDebug/debug.json* with content below is being ge
 {
   "buildInfo" : {
     "plugin" : {
-      "pluginVersion" : "6.0.2",
+      "pluginVersion" : "7.0.0-beta",
       "gradleVersion" : "5.0"
     },
     "gradle" : {
