+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tail/Log.kt
@@ -13,15 +13,16 @@ class Log(
 ) {
 
     val checksum = Formats.calculateChecksum(text)
+
     val message = messageLines.joinToString("\n")
+
     val messageChecksum = Formats.calculateChecksum(message)
 
     fun isError() = level == "ERROR"
 
     @Suppress("MagicNumber")
     fun isOlderThan(minutes: Long = 0, seconds: Long = 0, millis: Long = 0) =
-            LocalDateTime
-                    .now()
+            LocalDateTime.now()
                     .minusMinutes(minutes)
                     .minusSeconds(seconds)
                     .minusNanos(millis * 1000_000)
@@ -30,11 +31,17 @@ class Log(
     companion object {
 
         private const val TIMESTAMP = """(?<timestamp>[0-9]{2}\.[0-9]{2}\.[0-9]{4}\s[0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{3})"""
+
         private const val LEVEL = """\*(?<level>[A-Z]+)\*"""
+
         private const val SOURCE = """(?<source>\[.*\])"""
+
         private const val MESSAGE = """(?<message>.*)"""
+
         private const val LOG_PATTERN = "$TIMESTAMP\\s$LEVEL\\s$SOURCE\\s$MESSAGE"
+
         private const val DATE_TIME_FORMAT = "dd.MM.yyyy HH:mm:ss.SSS"
+
         private val DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern(DATE_TIME_FORMAT)
 
         fun create(logLines: List<String>): Log {
