+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tail/LogAnalyzer.kt
@@ -1,6 +1,5 @@
 package com.cognifide.gradle.aem.instance.tail
 
-import com.cognifide.gradle.aem.common.file.FileOperations
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.ObsoleteCoroutinesApi
 import kotlinx.coroutines.channels.Channel
@@ -24,6 +23,7 @@ class LogAnalyzer(
 ) {
 
     private val errorsChannel = Channel<Log>(Channel.UNLIMITED)
+
     private var aggregatedErrors = mutableListOf<Log>()
 
     init {
@@ -50,7 +50,10 @@ class LogAnalyzer(
     }
 
     private suspend fun checkIfErrorsCannonadeEnded() {
-        if (aggregatedErrors.isEmpty()) return
+        if (aggregatedErrors.isEmpty()) {
+            return
+        }
+
         if (aggregatedErrors.last().isOlderThan(millis = options.notificationDelay)) {
             notificationChannel.send(ProblematicLogs(instanceName, aggregatedErrors))
             aggregatedErrors = mutableListOf()
@@ -60,41 +63,4 @@ class LogAnalyzer(
     companion object {
         private const val CHECKING_FOR_ERROR_CANNONADE_TO_END_INTERVAL_IN_MILLIS = 100L
     }
-}
-
-class ProblematicLogs(val instanceName: String, val logs: List<Log>) {
-    val size = logs.size
-}
-
-class Blacklist(
-    private val filters: List<(Log) -> Boolean> = emptyList(),
-    private val blacklists: List<String> = emptyList()
-) {
-
-    private val parser = Parser()
-    private val blacklist = loadBlacklists()
-
-    private fun loadBlacklists(): Map<String, Log> {
-        return (loadDefaultBlacklists(TailOptions.BLACKLIST_FILES_DEFAULT) +
-                loadConfiguredBlacklists(blacklists))
-                .flatten().map { it.messageChecksum to it }.toMap()
-    }
-
-    private fun loadConfiguredBlacklists(blacklists: List<String>): List<List<Log>> {
-        return blacklists.map { logFile ->
-            FileOperations.fromPathOrClasspath(logFile) { reader ->
-                parser.parseLogs(reader)
-            }
-        }
-    }
-
-    private fun loadDefaultBlacklists(blacklists: List<String>): List<List<Log>> {
-        return blacklists.mapNotNull { logFile ->
-            FileOperations.optionalFromPathOrClasspath(logFile) { reader ->
-                parser.parseLogs(reader)
-            }
-        }
-    }
-
-    fun isBlacklisted(log: Log) = blacklist.containsKey(log.messageChecksum) || !filters.none { it(log) }
 }
