+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tail/LogDestination.kt
@@ -1,0 +1,5 @@
+package com.cognifide.gradle.aem.instance.tail
+
+interface LogDestination {
+    fun dump(logs: List<Log>)
+}
