+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tail/LogSource.kt
@@ -1,0 +1,8 @@
+package com.cognifide.gradle.aem.instance.tail
+
+import java.io.BufferedReader
+
+interface LogSource {
+
+    fun <T> readChunk(parser: (BufferedReader) -> List<T>): List<T>
+}
