+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tail/ProblematicLogs.kt
@@ -1,0 +1,6 @@
+package com.cognifide.gradle.aem.instance.tail
+
+class ProblematicLogs(val instanceName: String, val logs: List<Log>) {
+
+    val size = logs.size
+}
