+++ src/test/resources/com/cognifide/gradle/aem/test/compose/assembly/buildSrc/build.gradle.kts
@@ -4,6 +4,6 @@ repositories {
 }
 
 dependencies {
-    implementation("com.cognifide.gradle:aem-plugin:6.0.2")
+    implementation("com.cognifide.gradle:aem-plugin:7.0.0-beta")
     implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.10")
 }
