+++ src/test/resources/com/cognifide/gradle/aem/test/compose/bundle-and-content/buildSrc/build.gradle.kts
@@ -4,5 +4,5 @@ repositories {
 }
 
 dependencies {
-    implementation("com.cognifide.gradle:aem-plugin:6.0.2")
+    implementation("com.cognifide.gradle:aem-plugin:7.0.0-beta")
 }
