+++ src/test/resources/com/cognifide/gradle/aem/test/debug/additional/buildSrc/build.gradle.kts
@@ -5,5 +5,5 @@ repositories {
 }
 
 dependencies {
-    implementation("com.cognifide.gradle:aem-plugin:6.0.2")
+    implementation("com.cognifide.gradle:aem-plugin:7.0.0-beta")
 }
