+++ src/test/resources/com/cognifide/gradle/aem/test/debug/minimal/debug.json
@@ -1,7 +1,7 @@
 {
   "buildInfo" : {
     "plugin" : {
-      "pluginVersion" : "6.0.2",
+      "pluginVersion" : "7.0.0-beta",
       "gradleVersion" : "4.10"
     },
     "gradle" : {
